Add scratch for a tool for searching .toc files.
I'm not going to finish this, since I wanted to find the dependency
which triggers a file to be included. And this is simply not written to
any .toc-file. I should have had a deeper look at them first :-\

Anyway, this code may be useful for something, so I'm committing it.